Some Progress... again
I found out that a lot of my problems were stemming from the way I was issuing my API routes. Before I was module.exporting the api routes as a function, using app (express). I removed the function and put a regular module.export at the bottom then used router (express) instead of app. This improved the functionality and enabled me to post new workouts, but I still need the exercises in the new workouts to caste correctly. it seems a little off