Revert to previous functionality
I had an epiphany this morning after I woke up that some of my previous revisions made while trying to get the code to work had interfered with the planned structure of the application.  After reverting some old code, the add new workout button executes is it was conceptualized to do so.