STATS! Exercises in seeds deleted
I got the stats to populate, although as it stands it marks each workout as a "day" without factoring in chronological order, but I'm glad I got the data to populate on the chart. It's important to note  that having pre-made exercises in the workouts table interfered with the populate function, so I removed them from the seed. I hope that doesn't complicate things for testers.