2nd attempt make heroku ready
this is my second attempt to fix the paths src'ed in the html so theyll be recognized by Heroku. For some reason, every js file in the index.html works except for workout.js. I have no idea why. They all have the same syntax and are located in the same directory.